medvault_backend/requirements.txt
Django>=5.2
djangorestframework>=3.16
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.0
django-cors-headers>=4.0 